npc: Do not ignore errors when creating chains
Also, move ensureRulesAtTop to net/iptables.go
